Create Homework 1
This program calculates the price of an European call and put

What is the strike? 36.50

What is the spot? 39.19

Time to maturity in years? 1

What is the risk free rate? 0.012455

What is the volatility? 0.2177

What is the dividend? 0.02089
The call price is $ 4.52 and the put price is $1.67

This program calculates the price of an European call and put

What is the strike?